LayoutTests/accessibility/dialog-showModal-expected.txt

This tests that dialog.showModal() makes other elements inert.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Dialog is hidden
PASS backgroundAccessible() is true
Dialog is displaying
PASS backgroundAccessible() is false
Dialog is not displaying
PASS backgroundAccessible() is true
PASS backgroundAccessible() is true
PASS okButton.isIgnored is false
Dialog is displaying
PASS backgroundAccessible() is false
PASS okButton.isIgnored is false
Dialog is displaying and aria-hidden=true
PASS backgroundAccessible() is false
Dialog is displaying and aria-hidden=false
PASS backgroundAccessible() is false
Dialog is displaying and aria-modal=false
PASS backgroundAccessible() is false
Dialog is not displaying with opacity 0
PASS backgroundAccessible() is false
Dialog is displaying with opacity 1
PASS backgroundAccessible() is false
Dialog is removed from DOM
PASS backgroundAccessible() is true
PASS successfullyParsed is true

TEST COMPLETE
Other page content with a dummy focusable element

Display a dialog
